delete card functionality added for recipe creator, confirmation window before delete, one recipe deleted for test purpose

+++ react_atsiskaitymas/src/components/contexts/RecipesContext.tsx
@@ -47,6 +47,9 @@ const RecipesProvider = ({ children }: ChildrenProp) => {
     }
 
     const removeOneRecipe = (id: Recipe['id']) => {
+        const confirm = window.confirm("Do you want to delete it?");
+        if (!confirm) return;
+
         fetch(`http://localhost:8080/recipes/${id}`, {
             method: "DELETE"
         })
